math-solver-railway/python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sympy==1.12
numpy==1.26.2
scipy==1.11.4
matplotlib==3.8.2
Pillow==10.1.0
pytesseract==0.3.10
easyocr==1.7.1
pdf2image==1.16.3
PyPDF2==3.0.1
reportlab==4.0.8
pydantic==2.5.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
opencv-python-headless==4.8.1.78
